Enforce order for elements in unordered data structures (#33)
* Lint for order of set/dict elements

* Fix linting order

There isn't a great way to make these come in the order that would be
logical to humans. Given that this is how these will also be presented
at https://pypi.org/classifiers/, we should sort them the same as they
are there until we find a better way to order these.